+++ src/lib/components/SkjemaSteg/ny/SkjemaStegButtons.tsx
@@ -1,70 +1,43 @@
 import * as React from "react";
-import {useState} from "react";
 import {Button, Loader} from "@navikt/ds-react";
 import {useTranslation} from "react-i18next";
-import {DigisosValidationError, KortSkjemaPage, SkjemaPage} from "./SkjemaSteg";
-import {DigisosLanguageKey} from "../../../i18n";
 import {ArrowRightIcon} from "@navikt/aksel-icons";
-import {scrollToTop} from "../../../utils";
-import {logWarning} from "../../../log/loggerUtils.ts";
-import {useSkjemaNavigation} from "../useSkjemaNavigation.ts";
 import {SkjemaStegCancelButtons} from "./SkjemaStegCancelButtons.tsx";
+import {useState} from "react";
 
-interface SkjemaStegNavigasjonProps {
-    loading?: boolean;
-    onConfirm?: () => Promise<void>;
-    confirmTextKey?: DigisosLanguageKey;
-    page: SkjemaPage | KortSkjemaPage;
-    onRequestNavigation?: () => Promise<unknown>;
+interface SkjemaStegButtonsV2Props {
+    isFinalStep?: boolean;
+    onNext: () => Promise<any>;
+    onPrevious?: () => Promise<any>;
 }
 
-export const SkjemaStegButtons = ({
-    loading,
-    onConfirm,
-    confirmTextKey = "skjema.knapper.neste",
-    page,
-    onRequestNavigation,
-}: SkjemaStegNavigasjonProps) => {
+export const SkjemaStegButtons = ({onNext, onPrevious, isFinalStep}: SkjemaStegButtonsV2Props) => {
     const {t} = useTranslation("skjema");
-    const [sendSoknadPending, setSendSoknadPending] = useState<boolean>(false);
-    const {gotoPage} = useSkjemaNavigation(page);
+    const [isPending, setIsPending] = useState<boolean>(false);
 
-    const requestNavigation = async (page: number) => {
-        if (onConfirm) {
-            setSendSoknadPending(true);
-            await onConfirm();
-            setSendSoknadPending(false);
-        }
+    const nextButtonText = isFinalStep ? "skjema.knapper.send" : "skjema.knapper.neste";
+    const nextButtonIcon = isPending ? <Loader /> : <ArrowRightIcon />;
 
-        try {
-            if (onRequestNavigation !== undefined) await onRequestNavigation();
-            gotoPage(page);
-        } catch (e: any) {
-            scrollToTop();
-            if (!(e instanceof DigisosValidationError)) await logWarning(`Nektet navigering: ${e}`);
-        }
+    const onClickNext = async () => {
+        setIsPending(true);
+        await onNext();
+        setIsPending(false);
     };
 
     return (
         <div>
             <div className={"!mt-12 md:!mt-16 lg:!mt-24 !mb-8 lg:!mb-16 space-x-3"}>
-                <Button
-                    variant="secondary"
-                    onClick={async () => await requestNavigation(page - 1)}
-                    disabled={loading || page <= 1}
-                >
+                <Button variant="secondary" onClick={onPrevious} disabled={onPrevious === undefined}>
                     {t("skjema.knapper.forrige")}
-                    {loading && <Loader className={"ml-2"} />}
                 </Button>
                 <Button
                     variant="primary"
-                    onClick={async () => await requestNavigation(page + 1)}
-                    disabled={sendSoknadPending || loading}
-                    icon={<ArrowRightIcon />}
+                    onClick={onClickNext}
+                    disabled={isPending}
+                    icon={nextButtonIcon}
                     iconPosition={"right"}
                 >
-                    {t(confirmTextKey)}
-                    {(sendSoknadPending || loading) && <Loader className={"ml-2"} />}
+                    {t(nextButtonText)}
                 </Button>
             </div>
             <SkjemaStegCancelButtons />
