+++ src/sider/08-vedlegg/index.tsx
@@ -11,7 +11,7 @@ import {SkjemaStegTitle} from "../../lib/components/SkjemaSteg/ny/SkjemaStegTitl
 import {useTranslation} from "react-i18next";
 import {SkjemaStegStepperV2} from "../../lib/components/SkjemaSteg/ny/SkjemaStegStepperV2.tsx";
 import {useNavigate} from "react-router";
-import {SkjemaStegButtonsV2} from "../../lib/components/SkjemaSteg/ny/SkjemaStegButtonsV2.tsx";
+import {SkjemaStegButtons} from "../../lib/components/SkjemaSteg/ny/SkjemaStegButtons.tsx";
 
 export const OkonomiskeOpplysningerView = () => {
     const {bekreftet, isLoading, sorterte, grupper} = useOpplysninger();
@@ -48,7 +48,7 @@ export const OkonomiskeOpplysningerView = () => {
             ))}
             <SkjemaContent className={cx("pb-12")}>
                 <Gruppe gruppeKey={lastGroup} opplysninger={sorterte.filter((x) => x.gruppe === lastGroup)} />
-                <SkjemaStegButtonsV2 onPrevious={async () => navigate(`../7`)} onNext={async () => navigate("../9")} />
+                <SkjemaStegButtons onPrevious={async () => navigate(`../7`)} onNext={async () => navigate("../9")} />
             </SkjemaContent>
         </SkjemaSteg>
     );
